add `handleImportAtRules` to CLI
